refactor: separate recorders in different instances
Later, we will be adding the browser recorder as another `recorder-cli` instance.
