--- components/recorder-cli/default/.ordering
@@ -1,5 +1,0 @@
-process.mjs
-remote.mjs
-mocha.mjs
-jest.mjs
-index.mjs
