--- components/recorder-cli/default/mocha.test.mjs
@@ -1,48 +1,0 @@
-import { cwd } from "node:process";
-import { mochaHooks as hooks } from "../../../lib/node/mocha-hook.mjs";
-import "../../__fixture__.mjs";
-import { toDirectoryUrl } from "../../url/index.mjs";
-import { convertPathToFileUrl } from "../../path/index.mjs";
-import {
-  createConfiguration,
-  extendConfiguration,
-} from "../../configuration/index.mjs";
-import { record } from "./mocha.mjs";
-
-const home = toDirectoryUrl(convertPathToFileUrl(cwd()));
-
-record(
-  extendConfiguration(
-    createConfiguration(home),
-    {
-      recorder: "mocha",
-      hooks: {
-        cjs: false,
-        esm: false,
-        eval: false,
-        apply: false,
-        http: false,
-        mysql: false,
-        pg: false,
-        sqlite3: false,
-      },
-    },
-    "file:///w:/base/",
-  ),
-);
-
-const { beforeEach, afterEach } = hooks;
-
-beforeEach.call({
-  currentTest: {
-    parent: {
-      fullTitle: () => "full-title-1",
-    },
-  },
-});
-
-afterEach.call({
-  currentTest: {
-    state: "passed",
-  },
-});
