+++ components/recorder-cli/jest/.env
@@ -1,2 +1,1 @@
-test
 node
