+++ components/recorder-cli/jest/.ordering
@@ -1,0 +1,1 @@
+index.mjs
