+++ components/recorder-cli/mocha/.env
@@ -1,0 +1,1 @@
+node
