+++ components/recorder-cli/mocha/index.test.mjs
@@ -1,4 +1,5 @@
 import { cwd } from "node:process";
+import { mochaHooks as hooks } from "../../../lib/node/mocha-hook.mjs";
 import "../../__fixture__.mjs";
 import { toDirectoryUrl } from "../../url/index.mjs";
 import { convertPathToFileUrl } from "../../path/index.mjs";
@@ -6,16 +7,15 @@ import {
   createConfiguration,
   extendConfiguration,
 } from "../../configuration/index.mjs";
-import { record } from "./process.mjs";
+import { record } from "./index.mjs";
 
-const base = toDirectoryUrl(convertPathToFileUrl(cwd()));
+const home = toDirectoryUrl(convertPathToFileUrl(cwd()));
 
 record(
   extendConfiguration(
-    createConfiguration("file:///w:/home/"),
+    createConfiguration(home),
     {
-      processes: [{ regexp: "", enabled: true }],
-      recorder: "remote",
+      recorder: "mocha",
       hooks: {
         cjs: false,
         esm: false,
@@ -27,6 +27,22 @@ record(
         sqlite3: false,
       },
     },
-    base,
+    "file:///w:/base/",
   ),
 );
+
+const { beforeEach, afterEach } = hooks;
+
+beforeEach.call({
+  currentTest: {
+    parent: {
+      fullTitle: () => "full-title-1",
+    },
+  },
+});
+
+afterEach.call({
+  currentTest: {
+    state: "passed",
+  },
+});
