+++ components/recorder-cli/process/.env
@@ -1,0 +1,1 @@
+node
