+++ components/recorder-cli/process/index.mjs
