+++ components/recorder-cli/remote/.env
@@ -1,0 +1,1 @@
+node
