+++ lib/node/recorder.mjs
@@ -3,6 +3,8 @@ import { configuration } from "./configuration.mjs";
 export * from "./loader-esm.mjs";
 
 // Use top-level await to wait for ./configuration.mjs to update env variables.
-const { record } = await import("../../dist/bundles/recorder-cli.mjs");
+const { record } = await import(
+  `../../dist/bundles/recorder-${configuration.recorder}.mjs`
+);
 
 record(configuration);
